+++ packages/storefront/tests/unit/specs/generateVueMarkup.spec.ts
@@ -155,6 +155,31 @@ describe('generateVueControlledScript()', () => {
     `
     );
   });
+
+  it('should return a toggling eventHandler for toggleValue', () => {
+    const { states, eventHandler } = generateVueControlledScript(
+      'p-button',
+      [
+        [
+          'onClick',
+          {
+            target: 'p-popover',
+            prop: 'open',
+            toggleValue: true,
+          },
+        ],
+      ],
+      {}
+    );
+    expect(states).toMatchInlineSnapshot('""');
+    expect(eventHandler).toMatchInlineSnapshot(
+      `
+      "  const onClick = () => {
+          open.value = !open.value;
+        }"
+    `
+    );
+  });
 });
 
 describe('generateVueProperties()', () => {